Lesson 20 - Counting lines in parallel with manual thread management and using an Executor Service.

+++ weather-streams/src/test/java/pt/isel/leic/mpd/v1819/li41d/utils/ExecutionBenchmark.java
@@ -6,7 +6,7 @@ public class ExecutionBenchmark {
         runnable.run();
         final long end = System.currentTimeMillis();
 
-        System.out.println("Operation took " + (end-start)/1000 + "s");
+        System.out.println("Operation took " + (double)(end-start)/1000 + "s");
 
     }
